refactor user entity

+++ back_end/src/main/java/com/university/booking_university_project/modules/user/UserServiceImpl.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class UserServiceImpl extends BaseServiceImpl<User, Long, UserRepository> implements UserService {
+public class UserServiceImpl extends BaseServiceImpl<User, Integer, UserRepository> implements UserService {
 
     private final UserRepository userRepository;
